feat(cli): ngbp-cli is a thin wrapper for ngbp
Most of the code here (and all of the logic) has been moved to ngbp proper. The ngbp-cli binary is
now only responsible for ensuring that ngbp is installed locally and for calling it. All other
logic is upstream to prevent needing to update this very frequently.

More information at ngbp@40f3a8fcedd4723447ac9544d115eaffb4aca290.